Moved files into components folders. Added files for new future features, a progress bar and a way to set your daily fasting goal.
