+++ public/next.config.mjs
@@ -1,0 +1,7 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    output: 'export', // Outputs a Single-Page Application (SPA).
+    distDir: './build', // Changes the build output directory to `./dist`.
+  }
+   
+  export default nextConfig
