+++ src/App.js
@@ -1,4 +1,5 @@
-import Timer from "./Timer";
+import Timer from "./components/Timer";
+import Goals from "./components/Goals";
 import "./App.css";
 
 function App() {
@@ -7,6 +8,7 @@ function App() {
       <div style={{ textAlign: "center", padding: "50px" }}>
         <h1>Intermittent Fasting Timer</h1>
         <Timer />
+        <Goals />
       </div>
     </div>
   );
