+++ src/components/Goals/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Goals = () => {
+    return (
+        <div>
+            <p>Set goals</p>
+        </div>
+    )
+}
+
+export default Goals
