+++ src/components/ProgressBar/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ProgressBar = () => {
+
+    return (
+        
+            <h2>In progress</h2>
+    )
+}
+
+export default ProgressBar
